requirements.txt
streamlit
google.generativeai
import google.generativeai as genai
python-dotenv